move users and projects into separate files
