+++ frontend/src/components/admin/projects.tsx
@@ -1,0 +1,5 @@
+import React from "react"
+
+export default function Projects({ token }: { token: string }) {
+  return <div data-testid="projects-admin-widget">Projects</div>
+}
